Create UserRoleRepository

+++ core/src/main/java/ua/edu/chdtu/deanoffice/repository/ApplicationUserRepository.java
@@ -7,5 +7,5 @@ import ua.edu.chdtu.deanoffice.entity.ApplicationUser;
 @Repository
 public interface ApplicationUserRepository extends JpaRepository<ApplicationUser, Long> {
     ApplicationUser findByUsername(String username);
-    ApplicationUser findById(int id);
+    ApplicationUser findById(long id);
 }
